requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# Deep Learning
torch>=2.5.0
transformers>=4.35.2
timm>=0.9.0

# PDF Processing
pypdfium2>=4.30.0
img2table>=1.4.1

# Image Processing
pillow>=9.1

# Data
pandas>=2.3.0
tabulate>=0.9.0

# Visualization
matplotlib>=3.10.0

# Utils
typing_extensions>=4.9

# Testing
pytest>=9.0.0

# OCR (optionnel - installer selon le moteur choisi)
# Pour Tesseract: installer tesseract-ocr sur le système
# pip install pytesseract
# Pour PaddleOCR: pip install paddleocr
# Pour EasyOCR: pip install easyocr
